refactor: separate face plot funcs from class file
BREAKING CHANGE: call face plot funcs with a new namespace
